Make Admin Router postflight avoid using HTTP call for / on Agent AR.
Up until https://github.com/dcos/dcos/pull/2154/files Agent AR was
serving the root location (i.e. `/`), even thought it is incorrect and
it should be returning 404. Now that it has been fixed, a new endpoint
needs to be used for AR pooling which is suitable for both Open and EE,
Agent and Master. Such an endpoint seems to be
`/dcos-metadata/dcos-version.json`.